COnfigued the correct openshift domain suffix

+++ Docker/Production/entando-full-stack/test-all.sh
@@ -8,6 +8,12 @@ function cleanup(){
     docker volume rm entando-full-stack_entando-pg-volume &>/dev/null
     docker network rm entando-full-stack_entando-network &>/dev/null
 }
+if [ -n "$OPENSHIFT_REGISTRY" ]; then
+    INSTALLER_DIR=$(realpath $(dirname $(realpath ${BASH_SOURCE[0]}))/../../../Openshift/installers)
+    export TEST_DEPLOYMENT=true
+    export DESTROY_DEPLOYMENT=true
+    ${INSTALLER_DIR}/install-entando-full-stack.sh || { echo "Entando Full Stack Openshift test failed"; exit 1; }
+fi
 
 cleanup
 docker-compose -f docker-compose-qa.yml up -d || { echo "Could not bring down docker containers"; exit 1; }
@@ -45,11 +51,5 @@ docker run --rm --network=entando-full-stack_entando-network -e ENTANDO_APPBUILD
     || {  echo "The 'Login' test failed"; exit 1;  }
 cleanup
 
-if [ -n "$OPENSHIFT_REGISTRY" ]; then
-    INSTALLER_DIR=$(realpath $(dirname $(realpath ${BASH_SOURCE[0]}))/../../../Openshift/installers)
-    export TEST_DEPLOYMENT=true
-    export DESTROY_DEPLOYMENT=true
-    ${INSTALLER_DIR}/install-entando-full-stack.sh || { echo "Entando Full Stack Openshift test failed"; exit 1; }
-fi
 echo "Entando Full Stack tests successful"
 exit 0
